services/qwen2-vl/requirements.txt
fastapi[standard]
pydantic
qwen-vl-utils
# https://github.com/QwenLM/Qwen2-VL?tab=readme-ov-file#quickstart
git+https://github.com/huggingface/transformers
accelerate
pillow
torchvision